module refactor
Includes support for dynamically adding things like icons, labels, etc. All actions have been redone with the style of callbacks, so that it is easier to add actions and keep all code condensed. Other improvements made here and there. Socketio library updated to allow for easier updating of the front end as needed.